Fixed ClassRenamer and added MethodRenamer
The ClassRenamer had some issues with frames which have now been fixed.
The MethodRenamer should be fully functional but I wouldn't use it yet on big projects.
Still have to rework the FIeldRenamer.

+++ src/main/kotlin/xyz/xenondevs/obfuscator/util/CryptUtils.kt
@@ -1,10 +1,8 @@
 package xyz.xenondevs.obfuscator.util
 
-import kotlin.experimental.xor
-
 object CryptUtils {
 
     fun encrypt(text: String, key: String): String =
-            text.mapIndexed { index, c -> (c.toByte() xor key[index % key.length].toByte()).toChar() }.joinToString("")
+            text.mapIndexed { index, c -> (c.toInt() xor key[index % key.length].toInt()).toChar() }.joinToString("")
 
 }
